Make sure dialog appears on IE, Safari and Firefox
When you require('dialog-polyfill') the window.dialogPolyfill is
only available on Chrome.
In order to have the dialogPolyfill object availabe is using the
default from the require module.

Also the dialog should be under the body and should
not in nested divs
(see https://github.com/GoogleChrome/dialog-polyfill#limitations
for more details)