Add support for Google Tag Manager
In a similar way to the current support for Google Analytics (ENV variable `GA_TRACKING_ID`) if an ENV variable named `GTM_TRACKING_ID` is set, then the Google Tag Manager javascript snippet will be inserted.

These are not mutually exclusive, so it is still possible to set GA and GTM at the same time, although it defeats a bit the purpose of GTM as it is possible to set the GA tag from inside GTM anyways.

Re-position the snippet included block to be, as recommended, part of the head section. Also, included a block for the `noscript` version of the tracking, when javascript is not available (only for GTM as GA does not have an explicit noscript version).

More info:
https://developers.google.com/tag-manager/quickstart