Decrease default session duration
The `service.json` is the thing that defines the default session
duration used by the runner during start up, unless a specific form has
overridden the value in its' own `service.json`.

The timeout popup window also makes use of it to calculate the remaining
time notice for the user.