Allow custom button text
Users require the ability to add custom text to the green buttons on a
page. Button text is done at the global level via the Global Content.

Ideally we would like to be able to offer the ability for a user to edit
text at the button level on a specific page, but this proved to be
difficult.

Therefore in this PR we are adding 5 custom button definitions which can
be used instead of the usual choices.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>