Move remove anlaytics script tag
This script tag needs to be in the body of the page to have scope access
to the analytics object.

This is called if the user has rejected cookies. It is a backup just in
case of a race condition that for some reason prevented the analytics
cookies from being removed at the time the user clicked Reject.