Do not pass domain when removing cookie
Some anlaytics cookies use the hostname of the service as the value of
the 'Domain'. However some of them also use '.justice.gov.uk' as the
Domain.

We cannot grab the value of the Domain in the cookie from the client so
we have to do a basically attempt to delete all cookies. Not passing the
Domain should hopefully just remove everything.